Data/src/main/java/me/abdelrhman/data/store/ProjectsDataStoreFactory.kt
package me.abdelrhman.data.store

import me.abdelrhman.data.repository.ProjectsDataStore
import javax.inject.Inject

/**
 * Created by Abdelrhman
 * on 6/27/18.
 */

open class ProjectsDataStoreFactory @Inject constructor(
        private val projectsCacheDataStore: ProjectsCacheDataStore,
        private val projectsRemoteDataStore: ProjectsRemoteDataStore) {

    open fun getDataStore(projectsCached: Boolean,
                          cacheExpired: Boolean): ProjectsDataStore {
        return if (projectsCached && !cacheExpired) {
            projectsCacheDataStore
        } else {
            projectsRemoteDataStore
        }
    }

    open fun getCacheDataStore(): ProjectsDataStore {
        return projectsCacheDataStore
    }

    fun getRemoteDataStore(): ProjectsDataStore {
        return projectsRemoteDataStore
    }

}